add popup

+++ Frontend/src/app/components/edit-ship-navigation/edit-ship-navigation.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { CottageReservation } from 'src/app/interfaces/cottage-reservation';
 import { NavigationDto } from 'src/app/interfaces/navigation-dto';
 import { NavigationService } from 'src/app/services/NavigationService/navigation.service';
+import { ReservationService } from 'src/app/services/ReservationService/reservation.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-edit-ship-navigation',
@@ -15,12 +18,19 @@ export class EditShipNavigationComponent implements OnInit {
   navigations: NavigationDto[] = [];
   id: any;
   navigationn!: NavigationDto
-  constructor(private router: ActivatedRoute, private route: Router, private navigationService: NavigationService) {
+  haveReservations!: CottageReservation[]
+  constructor(private reservationService: ReservationService, private router: ActivatedRoute, private route: Router, private navigationService: NavigationService) {
     this.navigationn = {} as NavigationDto
   }
 
   ngOnInit(): void {
     this.findNavigation();
+    this.idShip = +this.router.snapshot.paramMap.get('id')!;
+
+    this.reservationService.getShipReservationById(this.idShip).subscribe((data) => {
+      this.haveReservations = data;
+
+    });
   }
 
 
@@ -34,32 +44,50 @@ export class EditShipNavigationComponent implements OnInit {
   }
 
   addNavigation() {
+    if (this.haveReservations.length != 0) {
+      Swal.fire({
+        icon: 'error',
+        title: 'Error',
+        text: 'The navigation cannot be added because ship has a reservation!',
+      })
 
-    this.navigationn.shipId = this.idShip;
+    }
+    else {
 
-    console.log(this.navigationn)
-    this.navigationService.addNavigation(this.navigationn).subscribe((data) => {
-      this.navigations = []
-      this.findNavigation();
+      this.navigationn.shipId = this.idShip;
 
-
-    });
-
-
-  }
-  deleteNavigation(id: any) {
-
-    this.navigationService.deleteNavigation(id, this.idShip)
-      .subscribe(data => {
+      console.log(this.navigationn)
+      this.navigationService.addNavigation(this.navigationn).subscribe((data) => {
         this.navigations = []
         this.findNavigation();
 
 
-
-
-
       });
+    }
 
+  }
+  deleteNavigation(id: any) {
+    if (this.haveReservations.length != 0) {
+      Swal.fire({
+        icon: 'error',
+        title: 'Error',
+        text: 'The navigation cannot be deleted because ship has a reservation!',
+      })
+
+    }
+    else {
+
+      this.navigationService.deleteNavigation(id, this.idShip)
+        .subscribe(data => {
+          this.navigations = []
+          this.findNavigation();
+
+
+
+
+
+        });
+    }
 
 
   }
