+++ Frontend/src/app/components/edit-ship-rules/edit-ship-rules.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { CottageReservation } from 'src/app/interfaces/cottage-reservation';
 import { RuleDto } from 'src/app/interfaces/rule-dto';
+import { ReservationService } from 'src/app/services/ReservationService/reservation.service';
 import { RuleService } from 'src/app/services/RuleService/rule.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-edit-ship-rules',
@@ -15,12 +18,18 @@ export class EditShipRulesComponent implements OnInit {
   rules: RuleDto[] = [];
   id: any;
   rulee!: RuleDto
+  haveReservations!: CottageReservation[]
 
-  constructor(private ruleService: RuleService, private router: ActivatedRoute, private route: Router) {
+  constructor(private ruleService: RuleService, private reservationService: ReservationService, private router: ActivatedRoute, private route: Router) {
     this.rulee = {} as RuleDto;
   }
 
   ngOnInit(): void {
+    this.idShip = +this.router.snapshot.paramMap.get('id')!;
+    this.reservationService.getShipReservationById(this.idShip).subscribe((data) => {
+      this.haveReservations = data;
+
+    });
 
 
     this.findRules();
@@ -34,33 +43,51 @@ export class EditShipRulesComponent implements OnInit {
   }
 
   addRule() {
+    if (this.haveReservations.length != 0) {
+      Swal.fire({
+        icon: 'error',
+        title: 'Error',
+        text: 'The rule cannot be changed because ship has a reservation!',
+      })
 
-    this.rulee.shipId = this.idShip;
+    }
+    else {
 
-    console.log(this.rulee)
-    this.ruleService.addRulebyShip(this.rulee).subscribe((data) => {
-      this.rules = []
-      this.findRules();
+      this.rulee.shipId = this.idShip;
 
-
-    });
-
-
-  }
-  deleteRule(id: any) {
-
-    this.ruleService.deleteRulebyShip(id, this.idShip)
-      .subscribe(data => {
+      console.log(this.rulee)
+      this.ruleService.addRulebyShip(this.rulee).subscribe((data) => {
         this.rules = []
         this.findRules();
 
 
-
-
-
       });
+    }
+
+  }
+  deleteRule(id: any) {
+    if (this.haveReservations.length != 0) {
+      Swal.fire({
+        icon: 'error',
+        title: 'Error',
+        text: 'The rule cannot be deleted because ship has a reservation!',
+      })
+
+    }
+    else {
+
+      this.ruleService.deleteRulebyShip(id, this.idShip)
+        .subscribe(data => {
+          this.rules = []
+          this.findRules();
 
 
+
+
+
+        });
+
+    }
 
   }
 }
